Ticket model improvements
