+++ src/lib/model/ticket/ticket-extra-option.ts
@@ -1,0 +1,5 @@
+export interface TicketExtraOption {
+  title: string;
+  value: string;
+  color?: string;
+}
