+++ src/lib/model/ticket/ticket-extra-text.ts
@@ -1,0 +1,5 @@
+export interface TicketExtraText {
+  title: string;
+  value: string;
+  url?: string;
+}
