+++ src/lib/model/ticket/ticket-extra-url.ts
@@ -1,0 +1,4 @@
+export interface TicketExtraUrl {
+  title: string;
+  url: string;
+}
