+++ src/public-api/model.ts
@@ -1,4 +1,4 @@
-export { TicketCreationParameters } from '../lib/model/ticket-creation-parameters.js';
+export { TicketCreationParameters } from '../lib/model/ticket/ticket-creation-parameters.js';
 export { GraphqlResponse } from '../lib/model/graphql-response.js';
 export { ConnectionStatus } from '../lib/model/connection-status.js';
 export { Desk } from '../lib/model/desk.js';
@@ -6,9 +6,13 @@ export { Device } from '../lib/model/device.js';
 export { Id } from '../lib/model/id.js';
 export { Line } from '../lib/model/line.js';
 export { Location } from '../lib/model/location.js';
-export { TicketExtra } from '../lib/model/ticket-extra.js';
-export { TicketLabel } from '../lib/model/ticket-label.js';
-export { Ticket, TicketStatus } from '../lib/model/ticket.js';
+export { TicketExtra } from '../lib/model/ticket/ticket-extra.js';
+export { TicketLabel } from '../lib/model/ticket/ticket-label.js';
+export { Ticket } from '../lib/model/ticket/ticket.js';
+export { TicketStatus } from '../lib/model/ticket/ticket-status.js';
+export { TicketExtraText } from '../lib/model/ticket/ticket-extra-text.js';
+export { TicketExtraUrl } from '../lib/model/ticket/ticket-extra-url.js';
+export { TicketExtraOption } from '../lib/model/ticket/ticket-extra-option.js';
 export { User } from '../lib/model/user.js';
 export { Webhook } from '../lib/model/webhook.js';
 export { SimpleError } from '../lib/model/errors/simple-error.js';
